solve one more scala problem

+++ scala/src/main/scala/kata/MultiplesOf3Or5.scala
@@ -1,5 +1,5 @@
 package kata
 
 object MultiplesOf3Or5 extends App {
-    def solution(number: Int): Int = (1 until number).filter(num => num % 3 == 0 || num % 5 == 0).sum
+  def solution(number: Int): Long = (1L until number).filter(num => num % 3 == 0 || num % 5 == 0).sum
 }
